ALL_INTEGRATIONS/requirements.txt
# Core Django packages
Django>=4.0.0
djangorestframework>=3.14.0
python-decouple>=3.6
dj-database-url>=1.0.0
whitenoise>=6.2.0
pymysql>=1.0.0

# Authentication and security
django-cors-headers>=3.13.0

# Data processing and analysis
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.7.0
scikit-learn>=1.2.0
joblib>=1.2.0

# Process mining
pm4py>=2.7.0
graphviz>=0.20

# Machine Learning / Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Utilities
python-dateutil>=2.8.0
pytz>=2021.3
Pillow>=9.0.0

# Server
gunicorn>=20.1.0
RapidFuzz==3.14.3
tabulate>=0.9.0
